split order filter component
