+++ src/app/admin/order_management/page.tsx
@@ -3,13 +3,14 @@
 import { useState, useEffect } from "react";
 import { request } from "@/lib/utils/request";
 import AdminPageHeader from "@/components/admin/common/admin_page_header";
-import { List, UserCircle, RefreshCw, Search, X } from "lucide-react";
+import { List, UserCircle, RefreshCw } from "lucide-react";
 import { useTranslation } from "@/i18n/i18n_context";
 import DataTable, { IDataTableColumn } from "@/components/common/data_table";
 import ConfirmModal from "@/components/common/confirm_modal";
 import SuccessNotification from "@/components/common/success_notification";
 import { formatDate } from "@/lib/utils/date";
 import { ORDER_STATUS } from "@/constants/status";
+import OrderFilter from "@/components/admin/order/order_filter";
 
 // Info: (20260625 - Julian) 訂單操作
 const ORDER_ACTIONS = {
@@ -27,25 +28,6 @@ const ORDER_API_ROUTES = {
   RETRY: (id: string) => `/api/v1/admin/orders/${id}/retry`,
   BATCH_REACTIVATE: "/api/v1/admin/orders/batch_reactivate",
 } as const;
-
-// Info: (20260625 - Julian) 訂單類型
-const ORDER_TYPE = {
-  BALANCE_SHEET: "balance_sheet", // Info: (20260625 - Julian) 資產負債表
-  CASH_FLOW: "cash_flow", // Info: (20260625 - Julian) 現金流量表
-  INCOME_STATEMENT: "income_statement", // Info: (20260625 - Julian) 損益表
-  IRSC: "irsc", // Info: (20260625 - Julian) 智能企業評級
-  FINANCIAL_COMPLIANCE: "financial_compliance", // Info: (20260625 - Julian) 財報合規性
-  FINANCIAL_HEALTH: "financial_health", // Info: (20260625 - Julian) 財務健康
-  MARKET_TRENDS: "market_trends", // Info: (20260625 - Julian) 市場趨勢
-  INDUSTRY_DEVELOPMENT: "industry_development", // Info: (20260625 - Julian) 產業趨勢
-  FINANCIAL_PRODUCT_RATING: "financial_product_rating", // Info: (20260625 - Julian) 投資評級
-  CARBON_HEALTH_CHECK: "carbon_health_check", // Info: (20260625 - Julian) 碳健檢
-  NET_ZERO_EMISSIONS: "net_zero_emissions", // Info: (20260625 - Julian) 碳中和
-  AI_CONSULTING: "ai_consulting", // Info: (20260625 - Julian) AI 諮詢
-  JOURNAL_UPLOAD: "journal_upload", // Info: (20260625 - Julian) 日記帳建檔
-  CERTIFICATE_ANALYSIS: "certificate_analysis", // Info: (20260625 - Julian) 憑證分析
-  TRANSPORTATION_CARBON_FOOTPRINT: "transportation_carbon_footprint", // Info: (20260625 - Julian) 運輸碳足跡
-};
 
 interface IOrderManagementData {
   id: string;
@@ -616,156 +598,31 @@ export default function OrderManagementPage() {
           subtitle={t("order_management.subtitle")}
         />
 
-        {/* Info: (20260624 - Julian) Filters and Search Panel */}
-        <div className="flex flex-col gap-4 rounded-xl border border-gray-200 bg-white p-4 shadow-sm">
-          <div className="grid grid-cols-1 items-end gap-2 md:grid-cols-5">
-            {/* Info: (20260624 - Julian) Search Input */}
-            <div className="col-span-1 flex flex-col gap-1 md:col-span-2">
-              <p className="text-xs font-medium text-slate-500">
-                {t("order_management.table.user")} &{" "}
-                {t("order_management.table.order_id")}
-              </p>
-              <div className="relative">
-                <span className="absolute inset-y-0 left-0 flex shrink-0 items-center pl-3 text-gray-400">
-                  <Search size={16} />
-                </span>
-                <input
-                  type="text"
-                  placeholder={t(
-                    "order_management.table.search_bar_placeholder",
-                  )}
-                  value={searchInput}
-                  onChange={(e) => setSearchInput(e.target.value)}
-                  className="w-full rounded-lg border border-gray-300 bg-white py-2 pr-10 pl-10 text-sm placeholder-gray-400 transition-colors focus:border-orange-500 focus:ring-1 focus:ring-orange-500 focus:outline-none"
-                />
-                {searchInput && (
-                  <button
-                    type="button"
-                    onClick={() => setSearchInput("")}
-                    className="absolute inset-y-0 right-0 flex shrink-0 items-center pr-3 text-gray-400 hover:text-gray-600"
-                  >
-                    <X size={16} />
-                  </button>
-                )}
-              </div>
-            </div>
-
-            {/* Info: (20260624 - Julian) Type Dropdowns */}
-            <div className="flex flex-col gap-1">
-              <p className="text-xs font-medium text-slate-500">
-                {t("order_management.table.type")}
-              </p>
-              <select
-                value={type}
-                onChange={handleTypeChange}
-                className="rounded-lg border border-gray-300 bg-white px-3 py-2 text-sm text-gray-700 transition-colors focus:border-orange-500 focus:ring-1 focus:ring-orange-500 focus:outline-none"
-              >
-                <option value="ALL">{t("common.all")}</option>
-                {Object.values(ORDER_TYPE).map((typeVal) => (
-                  <option key={typeVal} value={typeVal}>
-                    {getOrderTypeLabel(typeVal)}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            {/* Info: (20260624 - Julian) Order Status Dropdowns */}
-            <div className="flex flex-col gap-1">
-              <p className="text-xs font-medium text-slate-500">
-                {t("order_management.table.order_status")}
-              </p>
-              <select
-                value={orderStatus}
-                onChange={handleOrderStatusChange}
-                className="rounded-lg border border-gray-300 bg-white px-3 py-2 text-sm text-gray-700 transition-colors focus:border-orange-500 focus:ring-1 focus:ring-orange-500 focus:outline-none"
-              >
-                <option value="ALL">{t("common.all")}</option>
-                <option value={ORDER_STATUS.PENDING}>PENDING</option>
-                <option value={ORDER_STATUS.PAYING}>PAYING</option>
-                <option value={ORDER_STATUS.PAID}>PAID</option>
-                <option value={ORDER_STATUS.PAYMENT_FAILED}>
-                  PAYMENT_FAILED
-                </option>
-                <option value={ORDER_STATUS.EXECUTING}>
-                  {t("order_management.table.processing")}
-                </option>
-                <option value={ORDER_STATUS.COMPLETED}>
-                  {t("order_management.table.executed")}
-                </option>
-                <option value={ORDER_STATUS.FAILED}>
-                  {t("order_management.table.failed")}
-                </option>
-                <option value={ORDER_STATUS.MINT_FAILED}>MINT_FAILED</option>
-                <option value={ORDER_STATUS.CANCEL}>
-                  {t("common.cancel")}
-                </option>
-              </select>
-            </div>
-
-            {/* Info: (20260624 - Julian) Execution Status Dropdowns */}
-            <div className="flex flex-col gap-1">
-              <p className="text-xs font-medium text-slate-500">
-                {t("order_management.table.execution_status")}
-              </p>
-              <select
-                value={executionStatus}
-                onChange={handleExecutionStatusChange}
-                className="rounded-lg border border-gray-300 bg-white px-3 py-2 text-sm text-gray-700 transition-colors focus:border-orange-500 focus:ring-1 focus:ring-orange-500 focus:outline-none"
-              >
-                <option value="ALL">{t("common.all")}</option>
-                <option value="PENDING">
-                  {t("order_management.table.pending")}
-                </option>
-                <option value="EXECUTING">
-                  {t("order_management.table.processing")}
-                </option>
-                <option value="COMPLETED">
-                  {t("order_management.table.executed")}
-                </option>
-                <option value="FAILED">
-                  {t("order_management.table.failed")}
-                </option>
-                <option value="CANCEL">{t("common.cancel")}</option>
-              </select>
-            </div>
-          </div>
-
-          {/* Info: (20260624 - Julian) Reset & Batch Reactivate Button */}
-          <div className="flex items-center justify-end gap-2">
-            {(searchInput ||
+        <OrderFilter
+          searchInput={searchInput}
+          setSearchInput={setSearchInput}
+          type={type}
+          handleTypeChange={handleTypeChange}
+          orderStatus={orderStatus}
+          handleOrderStatusChange={handleOrderStatusChange}
+          executionStatus={executionStatus}
+          handleExecutionStatusChange={handleExecutionStatusChange}
+          hasActiveFilters={
+            !!(
+              searchInput ||
               type !== "ALL" ||
               executionStatus !== "ALL" ||
               orderStatus !== "ALL" ||
               sortBy !== "createdAt" ||
-              sortOrder !== "desc") && (
-              <button
-                type="button"
-                onClick={handleResetFilters}
-                className="inline-flex items-center justify-center gap-1 rounded-lg border border-gray-300 bg-white px-4 py-2 text-sm font-medium text-gray-700 transition-colors hover:bg-gray-50 focus:ring-2 focus:ring-orange-500 focus:ring-offset-2 focus:outline-none md:w-auto"
-              >
-                <X size={16} />
-                <p>{t("common.clear_filters")}</p>
-              </button>
-            )}
-
-            <button
-              type="button"
-              disabled={!hasReactivatableOrders && !isBatchMode}
-              onClick={handleToggleBatchMode}
-              className={`inline-flex items-center justify-center gap-1 rounded-lg px-4 py-2 text-sm font-medium transition-colors focus:ring-2 focus:ring-orange-500 focus:ring-offset-2 focus:outline-none ${
-                isBatchMode
-                  ? "border border-orange-300 bg-orange-100 text-orange-700 hover:bg-orange-200"
-                  : !hasReactivatableOrders
-                    ? "cursor-not-allowed border border-gray-200 bg-gray-100 text-gray-400"
-                    : "border border-transparent bg-orange-600 text-white hover:bg-orange-700"
-              }`}
-            >
-              {isBatchMode
-                ? t("order_management.table.cancel_batch")
-                : t("order_management.table.batch_reactivate")}
-            </button>
-          </div>
-        </div>
+              sortOrder !== "desc"
+            )
+          }
+          handleResetFilters={handleResetFilters}
+          isBatchMode={isBatchMode}
+          handleToggleBatchMode={handleToggleBatchMode}
+          hasReactivatableOrders={hasReactivatableOrders}
+          getOrderTypeLabel={getOrderTypeLabel}
+        />
 
         <div className="overflow-hidden rounded-xl border border-gray-200 bg-white shadow-sm">
           <DataTable<IOrderManagementData>
@@ -829,7 +686,7 @@ export default function OrderManagementPage() {
         onClose={() => setSuccessNotif({ isOpen: false, message: "" })}
       />
 
-      {/* Sleek Floating Batch Action Bar */}
+      {/* Info: (20260626 - Julian) 漂浮的批量操作欄 */}
       {selectedOrderIds.length > 0 && (
         <div className="animate-in fade-in slide-in-from-bottom-5 fixed bottom-8 left-1/2 z-50 flex -translate-x-1/2 items-center gap-5 rounded-2xl border border-slate-800 bg-slate-900 px-6 py-4 shadow-2xl transition-all duration-300">
           <div className="flex items-center gap-2 border-r border-slate-700 pr-5">
